pyproject.toml
[tool.poetry]
name = "passive_listening_classifiers"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "<3.9,>=3.7.1"  # require Python 3.8 for compatibility (3.7 might work too)
sclog = "*"
grpcio = "^1.35.0"
protobuf = "^3.14.0"
colorlog = "^4.7.2"
torch = "^1.7.1"
tensorflow = "2.3.1"
pandas = "^1.2.1"
spacy = "^2.3.5"
scikit-learn = "^0.23.1"
nltk = "^3.5"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz"}
fuzzywuzzy = "^0.18.0"
sentence-transformers = "^1.2.0"
truecase = "^0.0.12"
gdown = "^3.13.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pylint = "*"
mypy = "*"
black = "*"
rope = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
